refactor(pack): refactor xdsl bridgemode
