+++ packages/manager/apps/telecom/src/app/telecom/pack/xdsl/modem/index.js
@@ -6,6 +6,7 @@ import firewall from './firewall';
 import wifi from './wifi';
 import wifiConfig from './wifi/config';
 import XdslModemDhcpBdhcp from './router/dhcp/bdhcp';
+import bridgeMode from './bridgeMode';
 
 import component from './pack-xdsl-modem.component';
 import { PACK_XDSL_MODEM } from './pack-xdsl-modem.constant';
@@ -23,6 +24,7 @@ angular
     wifi,
     wifiConfig,
     XdslModemDhcpBdhcp,
+    bridgeMode,
   ])
   .component('packXdslModem', component)
   .config(routing)
